requirements.txt
flask==2.0.1
flask-limiter==2.4.0
flask-cors==3.0.10
python-dotenv==0.19.0
PyJWT==2.1.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
azure-identity==1.7.0
azure-keyvault-secrets==4.3.0
python-jose==3.3.0
joblib==1.3.1
gunicorn==21.2.0
python-json-logger==2.0.7
redis==4.6.0
requests==2.31.0
prometheus-client==0.17.1
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1